개인_AI_코드/CLIP_기반_이미지_유사도_검색/deploy_model/requirements.txt
fastapi
uvicorn
python-multipart

torch
transformers
sentence-transformers

faiss-cpu
numpy
Pillow
xmltodict
requests
matplotlib
sentencepiece
